+++ x/staking/keeper/grpc_query.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	"context"
+	"fmt"
 	"strings"
 
 	"github.com/cosmos/cosmos-sdk/types/errors"
@@ -10,6 +11,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
 
+	customgovtypes "github.com/KiraCore/sekai/x/gov/types"
 	"github.com/KiraCore/sekai/x/staking/types"
 
 	"google.golang.org/grpc/codes"
@@ -52,6 +54,8 @@ func (q Querier) ValidatorByMoniker(ctx context.Context, request *types.Validato
 
 // Validators implements the Query all validators gRPC method
 func (q Querier) Validators(ctx context.Context, request *types.ValidatorsRequest) (*types.ValidatorsResponse, error) {
+	c := sdk.UnwrapSDKContext(ctx)
+
 	if request == nil {
 		return nil, status.Error(codes.InvalidArgument, "empty request")
 	}
@@ -61,7 +65,7 @@ func (q Querier) Validators(ctx context.Context, request *types.ValidatorsReques
 		return nil, status.Errorf(codes.InvalidArgument, "invalid validator status %s", request.Status)
 	}
 
-	store := sdk.UnwrapSDKContext(ctx).KVStore(q.keeper.storeKey)
+	store := c.KVStore(q.keeper.storeKey)
 
 	var validators []types.QueryValidator
 
@@ -117,8 +121,16 @@ func (q Querier) Validators(ctx context.Context, request *types.ValidatorsReques
 		if request.All || accumulate {
 			validators = append(validators, validator)
 		}
-		return true, nil
+		return !request.All, nil
 	})
+
+	var actors []string
+	if request.All {
+		for _, actor := range q.keeper.govkeeper.GetNetworkActorsByAbsoluteWhitelistPermission(c, customgovtypes.PermClaimValidator) {
+			fmt.Println(actor.Address)
+			actors = append(actors, actor.Address.String())
+		}
+	}
 
 	if err != nil {
 		return nil, status.Error(codes.Internal, err.Error())
@@ -128,7 +140,7 @@ func (q Querier) Validators(ctx context.Context, request *types.ValidatorsReques
 		pageRes = nil
 	}
 
-	response := types.ValidatorsResponse{Validators: validators, Pagination: pageRes}
+	response := types.ValidatorsResponse{Validators: validators, Pagination: pageRes, Actors: actors}
 
 	return &response, nil
 }